requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aiohttp=3.8.3=pypi_0
aiosignal=1.3.1=pypi_0
async-timeout=4.0.2=pypi_0
attrs=22.2.0=pypi_0
black=22.3.0=pypi_0
ca-certificates=2022.12.7=h4653dfc_0
charset-normalizer=2.1.1=pypi_0
click=8.1.3=pypi_0
discord=2.1.0=pypi_0
discord-py=2.1.0=pypi_0
flake8=6.0.0=pypi_0
frozenlist=1.3.3=pypi_0
idna=3.4=pypi_0
isort=5.11.4=pypi_0
libcxx=14.0.6=h2692d47_0
libffi=3.4.2=h3422bc3_5
libsqlite=3.40.0=h76d750c_0
libzlib=1.2.13=h03a7124_4
mccabe=0.7.0=pypi_0
multidict=6.0.4=pypi_0
mypy=0.991=pypi_0
mypy-extensions=0.4.3=pypi_0
ncurses=6.3=h07bb92c_1
openssl=3.0.7=h03a7124_1
pathspec=0.10.3=pypi_0
pip=22.3.1=pyhd8ed1ab_0
platformdirs=2.6.2=pypi_0
pycodestyle=2.10.0=pypi_0
pydantic=1.10.4=pypi_0
pyflakes=3.0.1=pypi_0
python=3.8.12=hd949e87_3_cpython
python-dotenv=0.21.0=pypi_0
readline=8.1.2=h46ed386_0
setuptools=65.6.3=pyhd8ed1ab_0
sqlite=3.40.0=h2229b38_0
tk=8.6.12=he1e0b03_0
tomli=2.0.1=pypi_0
typing-extensions=4.4.0=pypi_0
wheel=0.38.4=pyhd8ed1ab_0
xz=5.2.6=h57fd34a_0
yarl=1.8.2=pypi_0
